+++ src/main/kotlin/com/insyncwithfoo/pyrightls/configuration/AllConfigurations.kt
@@ -1,5 +1,6 @@
 package com.insyncwithfoo.pyrightls.configuration
 
+import com.insyncwithfoo.pyrightls.configuration.application.LogLevel
 import org.jetbrains.annotations.SystemDependent
 import com.insyncwithfoo.pyrightls.configuration.application.Configurations as ApplicationConfigurations
 import com.insyncwithfoo.pyrightls.configuration.project.Configurations as ProjectConfigurations
@@ -15,6 +16,7 @@ internal infix fun ApplicationConfigurations.mergeWith(other: ProjectConfigurati
         hoverSupport = this.hoverSupport,
         completionSupport = this.completionSupport,
         goToDefinitionSupport = this.goToDefinitionSupport,
+        logLevel = this.logLevel,
         
         projectExecutable = other.projectExecutable,
         autoSuggestExecutable = other.autoSuggestExecutable
@@ -30,6 +32,7 @@ internal data class AllConfigurations(
     val hoverSupport: Boolean,
     val completionSupport: Boolean,
     val goToDefinitionSupport: Boolean,
+    val logLevel: LogLevel,
     
     val projectExecutable: @SystemDependent String?,
     val autoSuggestExecutable: Boolean