+++ src/main/kotlin/com/insyncwithfoo/pyrightls/configuration/application/Configurations.kt
@@ -1,6 +1,15 @@
 package com.insyncwithfoo.pyrightls.configuration.application
 
+import com.insyncwithfoo.pyrightls.message
 import com.intellij.openapi.components.BaseState
+
+
+internal enum class LogLevel(val label: String) {
+    ERROR(message("configurations.global.logLevel.error")),
+    WARNING(message("configurations.global.logLevel.warning")),
+    INFORMATION(message("configurations.global.logLevel.information")),
+    TRACE(message("configurations.global.logLevel.trace"));
+}
 
 
 internal class Configurations : BaseState() {
@@ -13,4 +22,5 @@ internal class Configurations : BaseState() {
     var hoverSupport by property(true)
     var completionSupport by property(false)
     var goToDefinitionSupport by property(false)
+    var logLevel by enum(LogLevel.INFORMATION)
 }